+++ swap-controller-api/src/lib.rs
@@ -193,8 +193,10 @@ pub trait AsbApi {
     #[method(name = "set_external_bitcoin_redeem_address")]
     async fn set_external_bitcoin_redeem_address(
         &self,
-        address: Option<String>,
+        address: String,
     ) -> Result<(), ErrorObjectOwned>;
+    #[method(name = "clear_external_bitcoin_redeem_address")]
+    async fn clear_external_bitcoin_redeem_address(&self) -> Result<(), ErrorObjectOwned>;
     #[method(name = "refresh_bitcoin_wallet")]
     async fn refresh_bitcoin_wallet(&self) -> Result<(), ErrorObjectOwned>;
     #[method(name = "get_current_quote")]